requirements.txt
absl-py==2.1.0
addict==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
appnope==0.1.3
asttokens==2.4.1
astunparse==1.6.3
autograd==1.8.0
autoray==0.8.0
basicsr==1.4.2
beautifulsoup4==4.12.2
blinker==1.9.0
bs4==0.0.1
cachetools==6.2.1
certifi==2025.8.3
charset-normalizer==3.3.2
choreographer==1.1.2
click==8.1.7
comm==0.2.1
contourpy==1.2.1
cycler==0.12.1
Cython==3.2.3
debugpy==1.8.0
decorator==5.1.1
diastatic-malt==2.15.2
dill==0.4.0
executing==2.0.1
facexlib==0.3.0
fastapi==0.129.0
filelock==3.13.1
filterpy==1.4.5
Flask==3.1.0
fonttools==4.51.0
fsspec==2023.12.2
future==1.0.0
gast==0.6.0
GeoAlchemy2==0.18.1
geopandas==1.1.2
gfpgan==1.3.8
grpcio==1.62.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.6.0
importlib_metadata==7.1.0
iniconfig==2.1.0
install==1.3.5
ipykernel==6.28.0
ipython==8.20.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
kaleido==0.2.1
kiwisolver==1.4.5
lazy_loader==0.4
llvmlite==0.42.0
lmdb==1.4.1
logistro==1.1.0
lxml==6.0.2
Markdown==3.6
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mlxtend==0.23.4
moviepy==2.2.1
mpmath==1.3.0
mysql-connector-python==9.1.0
narwhals==1.34.1
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
numba==0.59.1
numpy==1.26.3
opencv-python==4.9.0.80
orjson==3.11.3
osmnx==2.0.7
packaging==23.2
Panda3D==1.10.15
panda3d-gltf==1.3.0
panda3d-simplepbr==0.13.1
pandas==2.1.4
parso==0.8.3
pennylane==0.43.0
pennylane_lightning==0.43.0
pexpect==4.9.0
pgvector==0.4.2
pillow==10.2.0
platformdirs==4.1.0
plotly==6.3.1
pluggy==1.6.0
polyline==2.0.4
proglog==0.1.12
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.7
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.17.2
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
pyogrio==0.12.1
pyparsing==3.1.2
pypdf==6.7.0
pyperclip==1.11.0
pyphen==0.14.0
pyproj==3.7.2
pytest==8.4.2
pytest-timeout==2.4.0
python-dateutil==2.8.2
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
qiskit==2.2.3
qiskit-aer==0.17.2
regex==2023.12.25
requests==2.31.0
rich==14.3.2
rustworkx==0.17.1
safetensors==0.7.0
scikit-image==0.23.1
scikit-learn==1.7.2
scipy==1.13.0
scipy-openblas32==0.3.30.0.7
screeninfo==0.8.1
sentence-transformers==5.2.2
setuptools==69.5.0
shapely==2.1.2
shellingham==1.5.4
simplejson==3.20.2
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.46
sswg==4.3.0
stack-data==0.6.3
starlette==0.52.1
stevedore==5.5.0
sympy==1.14.0
tb-nightly==2.17.0a20240413
tensorboard-data-server==0.7.2
termcolor==3.2.0
textstat==0.7.3
threadpoolctl==3.6.0
tifffile==2024.2.12
tokenizers==0.22.2
tomli==2.0.1
tomlkit==0.13.3
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==5.1.0
typer==0.23.1
typer-slim==0.23.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2023.4
ucimlrepo==0.0.7
urllib3==2.1.0
ursina==8.3.0
uvicorn==0.40.0
vl-convert-python==1.8.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
yapf==0.40.2
zipp==3.18.1

# Data Update & Scheduling
schedule==1.2.0
